0.11.3

+++ src/Maximizer/Maximizer.java
@@ -33,8 +33,8 @@ public class Maximizer {
 	public void calculateAndStore() {
 
 		Main.calculateDPS();
-		
-		if(Double.isNaN(Main.raw.perSecond)) {
+
+		if (Double.isNaN(Main.raw.perSecond)) {
 			Main.stop = true;
 			return;
 		}
@@ -45,7 +45,7 @@ public class Maximizer {
 
 		String build1 = "[" + simulatedMods.get(0).name + "][" + simulatedMods.get(1).name + "][" + simulatedMods.get(2).name + "][" + simulatedMods.get(3).name + "]";
 		String build2 = "[" + simulatedMods.get(4).name + "][" + simulatedMods.get(5).name + "][" + simulatedMods.get(6).name + "][" + simulatedMods.get(7).name + "]";
-		
+
 		double DPS = Main.raw.perSecond;
 		for (TTKTarget target : Main.groupTargets) {
 			TTKs.add(target.name + " [" + target.currentLevel + "]");
@@ -64,7 +64,7 @@ public class Maximizer {
 		if (currentMaxTTK < bestTTK) {
 			bestTTK = currentMaxTTK;
 		}
-		
+
 		completedIterations++;
 		Main.progressBar.setValue((completedIterations * 100) / totalIterations);
 	}
@@ -79,7 +79,7 @@ public class Maximizer {
 			updateRanks();
 			calculateAndStore();
 		} else {
-			for (int i = 0; i < modCount-1; i++) {
+			for (int i = 0; i < modCount - 1; i++) {
 				simulatedMods.set(modsToChange.get(modSlot - 1), null);
 				updateMods();
 				simulatedMods.set(modsToChange.get(modSlot - 1), possibleMods.get(i));
@@ -87,7 +87,7 @@ public class Maximizer {
 			}
 		}
 	}
-	
+
 	public void updateMods() {
 		possibleMods = new Vector<Mod>();
 		for (Mod mod : Main.selectedWeapon.modInit.mods) {
@@ -98,10 +98,10 @@ public class Maximizer {
 			}
 		}
 	}
-	
+
 	public void updateRanks() {
 		simulatedRanks = new Vector<Integer>();
-		for(Mod mod : simulatedMods) {
+		for (Mod mod : simulatedMods) {
 			simulatedRanks.add(mod.ranks);
 		}
 	}
@@ -111,12 +111,12 @@ public class Maximizer {
 		int emptyMods = 0;
 		Main.stop = false;
 		completedIterations = 0;
-		bestTTK = 9999999;
+		bestTTK = Double.POSITIVE_INFINITY;
 		results.clear();
 		modsToChange.clear();
 		simulatedMods.clear();
 		simulatedRanks.clear();
-		
+
 		// Identify empty mod slots and count them
 		for (int i = 0; i < 8; i++) {
 			if (Main.selectedWeapon.selectedMods.get(i).equals("--")) {
@@ -136,57 +136,61 @@ public class Maximizer {
 		simulatedMods.add(Main.selectedWeapon.getModByName(Main.selectedWeapon.modSixPanel.getSelectedMod()));
 		simulatedMods.add(Main.selectedWeapon.getModByName(Main.selectedWeapon.modSevenPanel.getSelectedMod()));
 		simulatedMods.add(Main.selectedWeapon.getModByName(Main.selectedWeapon.modEightPanel.getSelectedMod()));
-		
+
 		// Do
 		thatThang(emptyMods);
 
-		if(Main.stop == true) {
+		if (Main.stop == true) {
 			Main.output.append("Maximizer stopped\n");
 			return;
 		}
-		
+
 		// For each target: find its fastest kill time, DPS, etc
-		// Find fasted kill time for each target
-		for (int k = 0; k < targets; k++) {
+
+		if (!Main.smartMax.isSelected()) { // These results will be incorrect if smartmax is enabled.
+
+			// Find fasted kill time for each target
+			for (int k = 0; k < targets; k++) {
+				for (int p = results.size() - 1; p > 0; p--) {
+					if (results.get(p).TTKs.get(1 + 2 * k) < results.get(p - 1).TTKs.get(1 + 2 * k)) {
+						Collections.swap(results, p, p - 1);
+					}
+				}
+				Main.output.append("::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::" + "\n");
+				Main.output.append("Build with the fasted kill time on " + results.get(0).TTKs.get(2 * k) + "\n");
+				Main.output.append(results.get(0).build1 + "\n");
+				Main.output.append(results.get(0).build2 + "\n");
+				Main.output.append("DPS: " + f.format(results.get(0).DPS) + "\n");
+				Main.output.append("Time: " + f.format(results.get(0).TTKs.get(1 + 2 * k)) + " seconds" + "\n");
+			}
+
+			// Find build with highest DPS
 			for (int p = results.size() - 1; p > 0; p--) {
-				if (results.get(p).TTKs.get(1 + 2 * k) < results.get(p - 1).TTKs.get(1 + 2 * k)) {
+				if (results.get(p).DPS > results.get(p - 1).DPS) {
 					Collections.swap(results, p, p - 1);
 				}
 			}
 			Main.output.append("::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::" + "\n");
-			Main.output.append("Build with the fasted kill time on " + results.get(0).TTKs.get(2 * k) + "\n");
+			Main.output.append("Build with the highest DPS" + "\n");
 			Main.output.append(results.get(0).build1 + "\n");
 			Main.output.append(results.get(0).build2 + "\n");
 			Main.output.append("DPS: " + f.format(results.get(0).DPS) + "\n");
-			Main.output.append("Time: " + f.format(results.get(0).TTKs.get(1 + 2 * k)) + " seconds" + "\n");
-		}
 
-		// Find build with highest DPS
-		for (int p = results.size() - 1; p > 0; p--) {
-			if (results.get(p).DPS > results.get(p - 1).DPS) {
-				Collections.swap(results, p, p - 1);
+			// Find build with lowest average kill time
+			for (int p = results.size() - 1; p > 0; p--) {
+				if (results.get(p).average < results.get(p - 1).average) {
+					Collections.swap(results, p, p - 1);
+				}
 			}
-		}
-		Main.output.append("::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::" + "\n");
-		Main.output.append("Build with the highest DPS" + "\n");
-		Main.output.append(results.get(0).build1 + "\n");
-		Main.output.append(results.get(0).build2 + "\n");
-		Main.output.append("DPS: " + f.format(results.get(0).DPS) + "\n");
-
-		// Find build with lowest average kill time
-		for (int p = results.size() - 1; p > 0; p--) {
-			if (results.get(p).average < results.get(p - 1).average) {
-				Collections.swap(results, p, p - 1);
+			Main.output.append("::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::" + "\n");
+			Main.output.append("Build with the Lowest average kill time on selected targets" + "\n");
+			Main.output.append(results.get(0).build1 + "\n");
+			Main.output.append(results.get(0).build2 + "\n");
+			Main.output.append("Average time to kill: " + f.format(results.get(0).average) + "\n");
+			Main.output.append("--------------------------Individual TTKs----------------------------" + "\n");
+			for (int k = 0; k < targets; k++) {
+				Main.output.append(results.get(0).TTKs.get(2 * k) + ": " + f.format(results.get(0).TTKs.get(1 + 2 * k)) + " seconds" + "\n");
 			}
-		}
-		Main.output.append("::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::" + "\n");
-		Main.output.append("Build with the Lowest average kill time on selected targets" + "\n");
-		Main.output.append(results.get(0).build1 + "\n");
-		Main.output.append(results.get(0).build2 + "\n");
-		Main.output.append("Average time to kill: " + f.format(results.get(0).average) + "\n");
-		Main.output.append("--------------------------Individual TTKs----------------------------" + "\n");
-		for (int k = 0; k < targets; k++) {
-			Main.output.append(results.get(0).TTKs.get(2 * k) + ": " + f.format(results.get(0).TTKs.get(1 + 2 * k)) + " seconds" + "\n");
 		}
 
 		// Find build with the lowest maximum kill time
@@ -196,7 +200,7 @@ public class Maximizer {
 			}
 		}
 		Main.output.append("::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::" + "\n");
-		Main.output.append("Build with the Lowest maximum kill time on selected targets" + "\n");
+		Main.output.append("Build with the Lowest maximum kill time on selected targets:" + "\n");
 		Main.output.append(results.get(0).build1 + "\n");
 		Main.output.append(results.get(0).build2 + "\n");
 		Main.output.append("Maximum kill time: " + f.format(results.get(0).minmax) + "\n");
@@ -208,33 +212,35 @@ public class Maximizer {
 		Main.output.append("::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::" + "\n");
 
 		// Make CSV File
-		maximizerResults = new File("MaximizerResults.csv");
-		try {
-			if (!maximizerResults.exists()) {
-				maximizerResults.createNewFile();
+		if (!Main.smartMax.isSelected()) {
+			maximizerResults = new File("MaximizerResults.csv");
+			try {
+				if (!maximizerResults.exists()) {
+					maximizerResults.createNewFile();
+				}
+				BufferedWriter writer = new BufferedWriter(new FileWriter(maximizerResults));
+				// Create column titles
+				String Spacing = "";
+				for (int i = 0; i < targets * 2; i++) {
+					Spacing += ",";
+				}
+				String columnTitles = "Build,DPS,TTKs" + Spacing + "Average TTK" + "," + "Maximum TTK";
+				writer.write(columnTitles + "\n");
+				// Fill Rows
+				for (int i = 0; i < results.size(); i++) {
+					String lineToWrite = results.get(i).build1 + results.get(i).build2 + "," + results.get(i).DPS + "," + results.get(i).TTKs + "," + results.get(i).average + "," + results.get(i).minmax;
+					writer.write(lineToWrite + "\n");
+				}
+				writer.flush();
+				writer.close();
+				Main.output.append("Full results stored in MaximizerResults.csv in the install folder\n");
+			} catch (Exception e) {
+				Main.output.append("Could not create MaximizerResults.csv\n");
 			}
-			BufferedWriter writer = new BufferedWriter(new FileWriter(maximizerResults));
-			// Create column titles
-			String Spacing = "";
-			for (int i = 0; i < targets * 2; i++) {
-				Spacing += ",";
-			}
-			String columnTitles = "Build,DPS,TTKs" + Spacing + "Average TTK" + "," + "Maximum TTK";
-			writer.write(columnTitles + "\n");
-			// Fill Rows
-			for (int i = 0; i < results.size(); i++) {
-				String lineToWrite = results.get(i).build1 + results.get(i).build2 + "," + results.get(i).DPS + "," + results.get(i).TTKs + "," + results.get(i).average + "," + results.get(i).minmax;
-				writer.write(lineToWrite + "\n");
-			}
-			writer.flush();
-			writer.close();
-			Main.output.append("Full results stored in MaximizerResults.csv in the install folder\n");
-		} catch (Exception e) {
-			Main.output.append("Could not create MaximizerResults.csv\n");
 		}
 	}
 
-	public static class TTKresult {
+	class TTKresult {
 		public String build1 = "";
 		public String build2 = "";
 		public double DPS = 0;
